Optional: implement the new Equality approach
