--- src/core-taggeds-optional/Optional/Optional.T/Optional.T.Equality.GetHashCode.cs
@@ -1,17 +1,0 @@
-﻿#nullable enable
-
-namespace System
-{
-    partial struct Optional<T>
-    {
-        public override int GetHashCode()
-            =>
-            hasValue
-                ? value is not null
-                    ? HashCode.Combine(EqualityContract, true, EqualityComparer.GetHashCode(value))
-                    : HashCode.Combine(EqualityContract, true)
-                : HashCode.Combine(EqualityContract);
-
-        private static Type EqualityContract => typeof(Optional<T>);
-    }
-}
