+++ src/core-taggeds-optional/Optional/Optional.T/Optional.T.Equality.Impl.Common.cs
@@ -6,6 +6,8 @@ namespace System
 {
     partial struct Optional<T>
     {
+        private static Type EqualityContract => typeof(Optional<T>);
+
         private static IEqualityComparer<T> EqualityComparer => EqualityComparer<T>.Default;
     }
 }
