+++ src/core-taggeds-optional/Optional/Optional.T/Optional.T.Equality.Impl.Equals.cs
@@ -1,0 +1,22 @@
+﻿#nullable enable
+
+namespace System
+{
+    partial struct Optional<T>
+    {
+        public bool Equals(Optional<T> other)
+        {
+            if (hasValue != other.hasValue)
+            {
+                return false;
+            }
+
+            if (hasValue)
+            {
+                return EqualityComparer.Equals(value, other.value);
+            }
+
+            return true;
+        }
+    }
+}
