+++ src/core-taggeds-optional/Optional/Optional.T/Optional.T.Equality.Wrap.cs
@@ -1,0 +1,24 @@
+﻿#nullable enable
+
+namespace System
+{
+    partial struct Optional<T>
+    {
+        public static bool Equals(Optional<T> left, Optional<T> right)
+            =>
+            left.Equals(right);
+
+        public static bool operator ==(Optional<T> left, Optional<T> right)
+            =>
+            left.Equals(right);
+
+        public static bool operator !=(Optional<T> left, Optional<T> right)
+            =>
+            left.Equals(right) is false;
+
+        public override bool Equals(object? obj)
+            =>
+            obj is Optional<T> other &&
+            left.Equals(right);
+    }
+}
